Copying over Parkers work, adding rollover, fixing bug

+++ x-pack/plugin/migrate/src/main/java/org/elasticsearch/xpack/migrate/action/ReindexDataStreamIndexTransportAction.java
@@ -4,12 +4,12 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-
 package org.elasticsearch.xpack.migrate.action;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.ElasticsearchException;
+import org.elasticsearch.ResourceAlreadyExistsException;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
 import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
@@ -21,8 +21,11 @@ import org.elasticsearch.action.support.IndicesOptions;
 import org.elasticsearch.action.support.SubscribableListener;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.client.internal.Client;
+import org.elasticsearch.cluster.block.ClusterBlockException;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
+import org.elasticsearch.cluster.metadata.MetadataCreateIndexService;
 import org.elasticsearch.cluster.service.ClusterService;
+import org.elasticsearch.common.settings.IndexScopedSettings;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.core.TimeValue;
 import org.elasticsearch.index.reindex.BulkByScrollResponse;
@@ -36,57 +39,28 @@ import org.elasticsearch.transport.TransportService;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicReference;
 
-public class TransportReindexDataStreamIndexAction extends HandledTransportAction<
+public class ReindexDataStreamIndexTransportAction extends HandledTransportAction<
     ReindexDataStreamIndexAction.Request,
     ReindexDataStreamIndexAction.Response> {
 
-    private static final Logger logger = LogManager.getLogger(TransportReindexDataStreamIndexAction.class);
+    private static final Logger logger = LogManager.getLogger(ReindexDataStreamIndexTransportAction.class);
 
-    // copied from
-    // https://github.com/elastic/kibana/blob/8a8363f02cc990732eb9cbb60cd388643a336bed/x-pack
-    // /plugins/upgrade_assistant/server/lib/reindexing/index_settings.ts#L155
-    private static final Set<String> DISALLOWED_SETTINGS = Set.of(
-        // Private ES settings
-        "index.allocation.existing_shards_allocator",
-        "index.history.uuid",
-        "index.merge.enabled",
-        "index.provided_name",
-        "index.resize.source.name",
-        "index.resize.source.uuid",
-        "index.routing.allocation.initial_recovery._id",
-        "index.search.throttled",
-        "index.source_only",
-        "index.shrink.source.name",
-        "index.shrink.source.uuid",
-        "index.verified_before_close",
-        "index.store.snapshot.repository_name",
-        "index.store.snapshot.snapshot_name",
-        "index.store.snapshot.snapshot_uuid",
-        "index.store.snapshot.index_name",
-        "index.store.snapshot.index_uuid",
-
-        // TODO verify set to correct value during index creation`
-        "index.creation_date",
-        "index.uuid",
-        "index.version.created",
-        "index.blocks.write",
-        "index.frozen",
-
-        // Deprecated in 9.0
-        "index.version.upgraded"
-    );
+    private static final Set<String> SETTINGS_TO_ADD_BACK = Set.of(IndexMetadata.SETTING_BLOCKS_WRITE, IndexMetadata.SETTING_READ_ONLY);
 
     private static final IndicesOptions IGNORE_MISSING_OPTIONS = IndicesOptions.fromOptions(true, true, false, false);
     private final ClusterService clusterService;
     private final Client client;
+    private final IndexScopedSettings indexScopedSettings;
 
     @Inject
-    public TransportReindexDataStreamIndexAction(
+    public ReindexDataStreamIndexTransportAction(
         TransportService transportService,
         ClusterService clusterService,
         ActionFilters actionFilters,
-        Client client
+        Client client,
+        IndexScopedSettings indexScopedSettings
     ) {
         super(
             ReindexDataStreamIndexAction.NAME,
@@ -98,6 +72,7 @@ public class TransportReindexDataStreamIndexAction extends HandledTransportActio
         );
         this.clusterService = clusterService;
         this.client = client;
+        this.indexScopedSettings = indexScopedSettings;
     }
 
     @Override
@@ -109,6 +84,7 @@ public class TransportReindexDataStreamIndexAction extends HandledTransportActio
         var sourceIndexName = request.getSourceIndex();
         var destIndexName = generateDestIndexName(sourceIndexName);
         IndexMetadata sourceIndex = clusterService.state().getMetadata().index(sourceIndexName);
+        Settings settingsBefore = sourceIndex.getSettings();
 
         if (sourceIndex.getCreationVersion().isLegacyIndexVersion() == false) {
             logger.warn(
@@ -118,25 +94,53 @@ public class TransportReindexDataStreamIndexAction extends HandledTransportActio
             );
         }
 
-        SubscribableListener.<AcknowledgedResponse>newForked(l -> setReadOnly(sourceIndexName, l))
-            .<AcknowledgedResponse>andThen(l -> deleteDestIfExists(destIndexName, l))
+        AtomicReference<BulkByScrollResponse> reindexResponse = new AtomicReference<>();
+        SubscribableListener.<AcknowledgedResponse>newForked(l -> setBlockWrites(sourceIndexName, l))
+            .<AcknowledgedResponse>andThen(l -> deleteDestIfExists(request.getDeleteDestIfExists(), destIndexName, l))
             .<CreateIndexResponse>andThen(l -> createIndex(sourceIndex, destIndexName, l))
             .<BulkByScrollResponse>andThen(l -> reindex(sourceIndexName, destIndexName, l))
-            .<AcknowledgedResponse>andThen(l -> updateSettings(sourceIndex, destIndexName, l))
-            // TODO handle searchable snapshots
-            .andThenApply(ignored -> new ReindexDataStreamIndexAction.Response(destIndexName))
+            .andThenApply(r -> {
+                reindexResponse.set(r);
+                return r;
+            })
+            .<AcknowledgedResponse>andThen(l -> updateSettings(settingsBefore, destIndexName, l))
+            .andThenApply(ignored -> new ReindexDataStreamIndexAction.Response(destIndexName, reindexResponse.get().getCreated()))
             .addListener(listener);
     }
 
-    private void setReadOnly(String sourceIndexName, ActionListener<AcknowledgedResponse> listener) {
+    private void setBlockWrites(String sourceIndexName, ActionListener<AcknowledgedResponse> listener) {
         logger.info("Setting read only on source index [{}]", sourceIndexName);
         final Settings readOnlySettings = Settings.builder().put(IndexMetadata.INDEX_BLOCKS_WRITE_SETTING.getKey(), true).build();
         var updateSettingsRequest = new UpdateSettingsRequest(readOnlySettings, sourceIndexName);
-        var errorMessage = String.format(Locale.ROOT, "Could not set read-only on source index [%s]", sourceIndexName);
-        client.admin().indices().updateSettings(updateSettingsRequest, failIfNotAcknowledged(listener, errorMessage));
+        client.admin().indices().updateSettings(updateSettingsRequest, new ActionListener<>() {
+            @Override
+            public void onResponse(AcknowledgedResponse response) {
+                if (response.isAcknowledged()) {
+                    listener.onResponse(null);
+                } else {
+                    var errorMessage = String.format(Locale.ROOT, "Could not set read-only on source index [%s]", sourceIndexName);
+                    listener.onFailure(new ElasticsearchException(errorMessage));
+                }
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                if (e instanceof ClusterBlockException || e.getCause() instanceof ClusterBlockException) {
+                    // It's fine if read-only is already set
+                    listener.onResponse(null);
+                } else {
+                    listener.onFailure(e);
+                }
+            }
+        });
     }
 
-    private void deleteDestIfExists(String destIndexName, ActionListener<AcknowledgedResponse> listener) {
+    private void deleteDestIfExists(boolean doDelete, String destIndexName, ActionListener<AcknowledgedResponse> listener) {
+        if (doDelete == false) {
+            listener.onResponse(null);
+            return;
+        }
+
         logger.info("Attempting to delete index [{}]", destIndexName);
         var deleteIndexRequest = new DeleteIndexRequest(destIndexName).indicesOptions(IGNORE_MISSING_OPTIONS)
             .masterNodeTimeout(TimeValue.MAX_VALUE);
@@ -154,8 +158,27 @@ public class TransportReindexDataStreamIndexAction extends HandledTransportActio
         Map<String, Object> mapping = sourceMapping != null ? sourceMapping.rawSourceAsMap() : Map.of();
         var createIndexRequest = new CreateIndexRequest(destIndexName).settings(settings).mapping(mapping);
 
-        var errorMessage = String.format(Locale.ROOT, "Could not create index [%s]", destIndexName);
-        client.admin().indices().create(createIndexRequest, failIfNotAcknowledged(listener, errorMessage));
+        client.admin().indices().create(createIndexRequest, new ActionListener<>() {
+            @Override
+            public void onResponse(CreateIndexResponse response) {
+                if (response.isAcknowledged()) {
+                    listener.onResponse(null);
+                } else {
+                    var errorMessage = String.format(Locale.ROOT, "Could not create index [%s]", destIndexName);
+                    listener.onFailure(new ElasticsearchException(errorMessage));
+                }
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                // If dest index already exists, just reindex into the existing one
+                if (e instanceof ResourceAlreadyExistsException || e.getCause() instanceof ResourceAlreadyExistsException) {
+                    listener.onResponse(null);
+                } else {
+                    listener.onFailure(e);
+                }
+            }
+        });
     }
 
     private void reindex(String sourceIndexName, String destIndexName, ActionListener<BulkByScrollResponse> listener) {
@@ -168,10 +191,10 @@ public class TransportReindexDataStreamIndexAction extends HandledTransportActio
         client.execute(ReindexAction.INSTANCE, reindexRequest, listener);
     }
 
-    private void updateSettings(IndexMetadata sourceIndex, String destIndexName, ActionListener<AcknowledgedResponse> listener) {
+    private void updateSettings(Settings settingsBefore, String destIndexName, ActionListener<AcknowledgedResponse> listener) {
         logger.info("Adding settings from source index that could not be added before reindex");
 
-        Settings postSettings = getPostSettings(sourceIndex);
+        Settings postSettings = getPostSettings(settingsBefore);
         if (postSettings.isEmpty()) {
             listener.onResponse(null);
             return;
@@ -182,19 +205,25 @@ public class TransportReindexDataStreamIndexAction extends HandledTransportActio
         client.admin().indices().updateSettings(updateSettingsRequest, failIfNotAcknowledged(listener, errorMessage));
     }
 
-    // Filter source index settings to subset of settings that can be included during reindex
-    // TODO does tsdb start/end time get set in create request if copied into settings from source index
+    // Filter source index settings to subset of settings that can be included during reindex.
+    // Similar to the settings filtering done when reindexing for upgrade in Kibana
+    // https://github.com/elastic/kibana/blob/8a8363f02cc990732eb9cbb60cd388643a336bed/x-pack
+    // /plugins/upgrade_assistant/server/lib/reindexing/index_settings.ts#L155
     private Settings getPreSettings(IndexMetadata sourceIndex) {
-        return sourceIndex.getSettings().filter(settingName -> DISALLOWED_SETTINGS.contains(settingName) == false);
+        // filter settings that will be added back later
+        var filtered = sourceIndex.getSettings().filter(settingName -> SETTINGS_TO_ADD_BACK.contains(settingName) == false);
+
+        // filter private and non-copyable settings
+        var builder = MetadataCreateIndexService.copySettingsFromSource(false, filtered, indexScopedSettings, Settings.builder());
+        return builder.build();
     }
 
-    private Settings getPostSettings(IndexMetadata sourceIndex) {
-        // TODO populate with real values
-        return Settings.EMPTY;
+    private Settings getPostSettings(Settings settingsBefore) {
+        return settingsBefore.filter(SETTINGS_TO_ADD_BACK::contains);
     }
 
     public static String generateDestIndexName(String sourceIndex) {
-        return "upgrade-" + sourceIndex;
+        return "migrated-" + sourceIndex;
     }
 
     private static <U extends AcknowledgedResponse> ActionListener<U> failIfNotAcknowledged(
